feat: Add initial TypeScript definitions and configuration files
