+++ ngui/ui/src/components/forms/ConnectCloudAccountForm/FormElements/AwsConnectionForm/AwsConnectionForm.constants.tsx
@@ -1,4 +1,7 @@
-import { AWS_CNR } from "utils/constants";
+import React from "react";
+import { DOCS_HYSTAX_CONNECT_AWS_ROOT } from "urls";
+import { AWS_CNR, CONNECTION_TYPES } from "utils/constants";
+import { AwsTypeDescription } from "./AwsConnectionFormElements";
 
 export const AUTHENTICATION_TYPES = Object.freeze({
   ASSUMED_ROLE: "assumedRole",
@@ -13,3 +16,60 @@ export const authenticationTypes = [
   },
   { authenticationType: AUTHENTICATION_TYPES.ACCESS_KEY, messageId: "accessKey", cloudType: AWS_CNR }
 ];
+
+export const AWS_ROOT_INPUTS_FIELD_NAMES = {
+  IS_FIND_REPORT: "isFindReport",
+  CONFIG_SCHEME: "configScheme"
+};
+
+const awsDefaultAssumedRoleDescriptions = [
+  <AwsTypeDescription key="1" messageId="createAwsMemberAssumedRoleDescriptions" />,
+  <AwsTypeDescription
+    key="2"
+    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
+    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
+  />
+];
+
+const awsDefaultAccessKeyDescriptions = [
+  <AwsTypeDescription
+    key="1"
+    messageId="createAwsDefaultAssumedRoleDDocumentationReference1"
+    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
+    linkDisplayBlock
+  />,
+  <AwsTypeDescription key="2" messageId="createAwsDefaultAssumedRoleDDocumentationReference2" />,
+  <AwsTypeDescription
+    key="3"
+    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
+    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
+  />
+];
+
+const awsMemberAccessKeyDescriptions = [
+  <AwsTypeDescription
+    key="1"
+    messageId="createAwsDefaultAssumedRoleDDocumentationReference1"
+    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
+    linkDisplayBlock
+  />,
+  <AwsTypeDescription key="2" messageId="createAwsDefaultAssumedRoleDDocumentationReference4" />,
+  <AwsTypeDescription
+    key="3"
+    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
+    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
+    linkDisplayBlock={false}
+  />
+];
+
+export const awsConnectionAssumedRoleDescriptions = {
+  [CONNECTION_TYPES.AWS_MANAGEMENT]: awsDefaultAssumedRoleDescriptions,
+  [CONNECTION_TYPES.AWS_MEMBER]: awsDefaultAssumedRoleDescriptions,
+  [CONNECTION_TYPES.AWS_STANDALONE]: awsDefaultAssumedRoleDescriptions
+};
+
+export const awsConnectionKeyAccessDescriptions = {
+  [CONNECTION_TYPES.AWS_MANAGEMENT]: awsDefaultAccessKeyDescriptions,
+  [CONNECTION_TYPES.AWS_MEMBER]: awsMemberAccessKeyDescriptions,
+  [CONNECTION_TYPES.AWS_STANDALONE]: awsDefaultAccessKeyDescriptions
+};
