--- ngui/ui/src/components/forms/ConnectCloudAccountForm/FormElements/AwsConnectionForm/AwsConnectionForm.tsx
@@ -1,126 +1,0 @@
-import React, { ReactNode, useState } from "react";
-import Alert from "@mui/material/Alert";
-import Box from "@mui/material/Box";
-import Link from "@mui/material/Link";
-import Typography from "@mui/material/Typography";
-import { FormattedMessage } from "react-intl";
-import { DOCS_HYSTAX_CONNECT_AWS_ROOT } from "urls";
-import { CONNECTION_TYPES } from "utils/constants";
-import ButtonGroup from "../../../../ButtonGroup";
-import { AUTHENTICATION_TYPES, authenticationTypes } from "./AwsConnectionForm.constants";
-import useStyles from "./AwsConnectionForm.styles";
-import { AuthenticationType, AuthenticationTypeSelectorType, AwsTypeDescriptionProps } from "./AwsConnectionForm.types";
-
-const AwsTypeDescription = ({ messageId, linkUrl, linkDisplayBlock = false }: AwsTypeDescriptionProps) => {
-  const { classes } = useStyles();
-
-  return (
-    <div key={messageId} className={classes.description}>
-      <FormattedMessage
-        id={messageId}
-        values={{
-          link: (chunks: ReactNode[]) => {
-            const linkText = chunks.length > 0 ? chunks : linkUrl;
-            return (
-              <Link
-                data-test-id="link_guide"
-                style={{ display: linkDisplayBlock ? classes.linkInline : classes.linkBlock }}
-                href={linkUrl}
-                target="_blank"
-                rel="noopener"
-              >
-                {linkText}
-              </Link>
-            );
-          },
-          strong: (chunks: ReactNode) => <strong>{chunks}</strong>,
-          warning: (chunks: ReactNode) => <Alert severity="warning">{chunks}</Alert>
-        }}
-      />
-    </div>
-  );
-};
-
-export const AuthenticationTypeSelector = ({ authenticationType, setAuthenticationType }: AuthenticationTypeSelectorType) => {
-  const { classes } = useStyles();
-
-  return (
-    <Box alignItems="center" display="flex">
-      <Typography className={classes.authenticationLabel}>
-        <FormattedMessage id="authentication" />{" "}
-      </Typography>
-      <ButtonGroup
-        buttons={authenticationTypes.map((subtype) => ({
-          id: subtype.authenticationType,
-          messageId: subtype.messageId,
-          action: () => setAuthenticationType(subtype.authenticationType)
-        }))}
-        activeButtonId={authenticationType}
-        activeButtonIndex={undefined}
-        fullWidth={false}
-      />
-    </Box>
-  );
-};
-
-const awsDefaultAssumedRoleDescriptions = [
-  <AwsTypeDescription key="1" messageId="createAwsMemberAssumedRoleDescriptions" />,
-  <AwsTypeDescription
-    key="2"
-    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
-    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
-  />
-];
-
-const awsDefaultAccessKeyDescriptions = [
-  <AwsTypeDescription
-    key="1"
-    messageId="createAwsDefaultAssumedRoleDDocumentationReference1"
-    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
-    linkDisplayBlock
-  />,
-  <AwsTypeDescription key="2" messageId="createAwsDefaultAssumedRoleDDocumentationReference2" />,
-  <AwsTypeDescription
-    key="3"
-    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
-    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
-  />
-];
-
-const awsMemberAccessKeyDescriptions = [
-  <AwsTypeDescription
-    key="1"
-    messageId="createAwsDefaultAssumedRoleDDocumentationReference1"
-    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
-    linkDisplayBlock
-  />,
-  <AwsTypeDescription key="2" messageId="createAwsDefaultAssumedRoleDDocumentationReference4" />,
-  <AwsTypeDescription
-    key="3"
-    messageId="createAwsDefaultAssumedRoleDDocumentationReference3"
-    linkUrl={DOCS_HYSTAX_CONNECT_AWS_ROOT}
-    linkDisplayBlock={false}
-  />
-];
-
-export const awsConnectionAssumedRoleTypeDescriptions = {
-  [CONNECTION_TYPES.AWS_MANAGEMENT]: awsDefaultAssumedRoleDescriptions,
-  [CONNECTION_TYPES.AWS_MEMBER]: awsDefaultAssumedRoleDescriptions,
-  [CONNECTION_TYPES.AWS_STANDALONE]: awsDefaultAssumedRoleDescriptions
-};
-
-export const awsConnectionKeyAccessTypeDescriptions = {
-  [CONNECTION_TYPES.AWS_MANAGEMENT]: awsDefaultAccessKeyDescriptions,
-  [CONNECTION_TYPES.AWS_MEMBER]: awsMemberAccessKeyDescriptions,
-  [CONNECTION_TYPES.AWS_STANDALONE]: awsDefaultAccessKeyDescriptions
-};
-
-export const useAuthenticationType = () => {
-  const [authenticationType, setAuthenticationType] = useState<AuthenticationType>(() => AUTHENTICATION_TYPES.ASSUMED_ROLE);
-  return { authenticationType, setAuthenticationType };
-};
-
-export const getAwsConnectionTypeDescriptions = (authenticationType: AuthenticationType) =>
-  authenticationType === AUTHENTICATION_TYPES.ASSUMED_ROLE
-    ? awsConnectionAssumedRoleTypeDescriptions
-    : awsConnectionKeyAccessTypeDescriptions;
